Idiomatic shuffle
