+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "proto"]
+	path = proto
+	url = https://github.com/gitops-ci-cd/proto.git
