+++ src/controllers/errorController.js
@@ -1,0 +1,4 @@
+export const simulateError = (req, res, next) => {
+  const error = new Error("Test error");
+  next(error);
+};
