+++ src/controllers/greetingController.js
@@ -1,0 +1,4 @@
+export const greet = (req, res) => {
+  const audience = req.query.audience || "World";
+  res.send(`Hello, ${audience}!`);
+};
