+++ src/index.js
@@ -1,40 +1,7 @@
-// Import Express
-import express, { json } from "express";
+import app from "./app.js";
 
-// Create an instance of an Express app
-const app = express();
+const PORT = process.env.PORT || 8080;
 
-// Middleware to parse JSON bodies
-app.use(json());
-
-// Define a simple route
-app.get("/hello", (req, res) => {
-  const audience = req.query.audience || "World";
-
-  res.send(`Hello, ${audience}!`);
+app.listen(PORT, () => {
+  console.log(`Server is running on port ${PORT}`);
 });
-
-app.get("/salutations", (req, res) => {
-  const audience = req.query.audience || "World";
-
-  res.send(`Yo, ${audience}!`);
-});
-
-// Define a route to simulate an error
-app.get("/boom", (_req, _res, next) => {
-  const error = new Error("Test error");
-
-  next(error);
-});
-
-// Middleware to handle 404 - Not Found errors
-app.use((_req, res) => {
-  res.status(404).send("Route not found");
-});
-
-// Middleware to handle errors
-app.use((_err, _req, res, _next) => {
-  res.status(500).send("Something went wrong!");
-});
-
-export default app; // Export the app for testing
