+++ src/middleware/errorMiddleware.js
@@ -1,0 +1,4 @@
+export const errorHandler = (err, req, res, _next) => {
+  console.error(err.stack);
+  res.status(500).send("Something went wrong!");
+};
