+++ src/middleware/index.js
@@ -1,0 +1,7 @@
+import { notFoundHandler } from "./notFoundMiddleware.js";
+import { errorHandler } from "./errorMiddleware.js";
+
+export default [
+  notFoundHandler,
+  errorHandler,
+];
