+++ src/middleware/notFoundMiddleware.js
@@ -1,0 +1,3 @@
+export const notFoundHandler = (req, res) => {
+  res.status(404).send("Route not found");
+};
