+++ src/routes/errorRoutes.js
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { simulateError } from "../controllers/errorController.js";
+
+const router = Router();
+
+router.get("/boom", simulateError);
+
+export default router;
