+++ src/routes/greetingRoutes.js
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { greet } from "../controllers/greetingController.js";
+
+const router = Router();
+
+router.get("/hello", greet);
+
+export default router;
