+++ src/routes/index.js
@@ -1,0 +1,11 @@
+import { Router } from "express";
+import greetingRoutes from "./greetingRoutes.js";
+import errorRoutes from "./errorRoutes.js";
+
+const router = Router();
+
+// Mount routes
+router.use(greetingRoutes);
+router.use(errorRoutes);
+
+export default router;
