--- src/server.js
@@ -1,7 +1,0 @@
-import app from "./index.js";
-
-const PORT = process.env.PORT || 3000;
-
-app.listen(PORT, () => {
-  console.log(`Server is running on port ${PORT}`);
-});
